added test to reveal bug reported by John Farrell
quote:
We are seeing a problem in Jacorb Notification in class EqOperator in method
evaluate(EvaluationContext), where _left is null so the .compareTo(_right) is
failing with a NullPointerException (line 46). I believe this is caused by
using an ETCL expression where I am saying "x == 3", when "x" is not defined
in the event.