this is a fix for #645
if the specified port param lies within the port range of the Factory
it first tries to open this port. If this fails another port from the port range
is tried.

i have also refactored PortRangeServerSocketFactory to delegate the actual creation to an instance of DefaultServerSocketFactory and added a testcase.