RT#1056: Codeset issue
the problem occurs due to the fact that System.getProperties returns a
different value depending on *both* the value of the native encoding
(e.g. LC_ALL=UTF8) and the operating system.

For instance for Linux it may return ISO-8859-1 while for HP it returns
8859_1. This may be shown by running CodeSet.main(). We are going
to modify how it retrieves the encoding name so it uses the canonical
name (ISO8859_1 on linux, hp and solaris).

Modified:
CodeSet.java to use canonical name
PrintIOR.java to output encoding details if in debug mode and tidied
output format up a bit