Compute the size of fragmented responses
When `GIOPConnection` returns a response built from multiple fragments,
it sets its size to the size of the first fragment.

This patch fixes this behavior so that the size of combined response is
computed correctly (as sum of fragments' sizes). As a result, if an AMI
CORBA call happens to result in an exception (returned in fragments),
`ReplyInputStream.getBody()` call inside `ExceptionHolderImpl`
constructor works correctly, instead of throwing
`NegativeArraySizeException`.