[Fix for bug 1014]
A single instance of 'org.jacorb.orb.Delegate' can be used by multiple
threads to perform invocations concurrently, which are sent using a
'ClientConnection' stored in an attribute of 'Delegate'. Eventually this
'ClientConnection' can be replaced due to the processing of a
LocationForward (e.g. replied by the ImR) or a ForwardRequest (thrown by
ClientInterceptors). In such cases, the replaced 'ClientConnection' is
closed if it is not used by any other 'Delegate'. However if one thread
is still waiting for a reply while the connection is replaced and closed
(no clients left!) the pending requests will result in a incorrect
'COMM_FAILURE'.

Another similar scenario, but which is much more difficult to reproduce
in a test case, is that one 'ClientConnection' might be replaced and
closed just after another thread gets a reference to the connection and
is about to write on the connection. In such case, when the closed
connection is used it automatically reconnects and but does not
processes any incoming messages (due to 'GIOPConnection.do_close ==
true'), therefore the request is never answered and the calling thread
freezes.

To avoid all these problems related to an concurrent close of a
'ClientConnection' while it is being used by other threads (to read
replies or send requests) I tried to make that whenever a
'ClientConnection' is obtained to perform a request, its client
reference count (ClientConnection.incClients) is increased. When the
request is finally completed, the client couter is decreased
(ClientConnection.decClients). This should avoid that the
'ClientConnection' be closed while in use to make a request.

I also added a test case for this bug 1014 and had to change test case
for bug 1012 because it was broken by this proposed solution. The test
case for bug 1012 seems to be a copy of the code provided to show some
of the problems reported in bug 1012. Because of that I suspect such
test case enforces some of the wrong behaviour (an extra call) that were
intended to be reported in bug 1012. For this reason I suggest that
someone take a better look at test case for bug 1012 or remove it from
the test suite.