Issue unmarshalling user exception with Any
During an unmarshal of a UserException with an Any type we ran into a
marshal exception.
It looks like the root cause is during the insert in
handle_receive_exception, we do a reflection call which calls ends up
calling <MyUserException>Helper.read on the stream.
If the <MyUserException> contains any corba types, we need to check
the typecode which in DelegatingTypeCodeReader.readTypeCode we add a
mark to the input stream. When we later call a reset in
handle_receive_exception, it resets to the typecode instead of the start
of the exception.  Finally in the generated
<ServerType>Stub.remoteMethod, in the catch block for the
ApplicationException, we call <MyUserException>Helper.read on the input
steam.  Because the stream is not at the start of the exception we get a
marshal error.
This change saves the position of the start of the exception in
handle_receive_exception so we return to the start of the exception
instead of the last mark in the input stream.