Add a test for exceptions with corba any variables
Prior to the changes in "Issue unmarshalling user exception with Any"
during the test: test_relative_roundtrip_fwdcall_at_rec_exc_with_any_OK
we would get a marshal exception indicating the wrong id.  The code
expects to see the id type name of the exception, but the input stream
was reset to the location of the any variable due to the stream getting
marked while reading the valuetype.